requirements/requirements-gpu.txt
tqdm==4.15.0
tflearn==0.3.2
numpy==1.16.0
protobuf==3.19.6
tensorflow-gpu==1.15.2
pytest==3.6.0
jsonlines==1.2.0
seaborn==0.9.0
black
pylint